requirements.txt
discord.py==2.3.2
aiohttp==3.9.1
beautifulsoup4==4.12.2
python-dotenv==1.0.0
pyyaml==6.0.1
requests==2.31.0